# secp256k1: Optimize k splitting with mod n scalar.